+++ daemons/based/cibmon.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2004-2019 the Pacemaker project contributors
+ * Copyright 2004-2020 the Pacemaker project contributors
  *
  * The version control history for this file may have further details.
  *
@@ -28,7 +28,6 @@
 
 #include <crm/common/ipc.h>
 #include <crm/pengine/status.h>
-#include <../lib/pengine/unpack.h>
 
 #include <crm/cib.h>
