+++ lib/pengine/pe_status_private.h
@@ -1,5 +1,7 @@
 /*
- * Copyright 2018 Andrew Beekhof <andrew@beekhof.net>
+ * Copyright 2018-2020 the Pacemaker project contributors
+ *
+ * The version control history for this file may have further details.
  *
  * This source code is licensed under the GNU Lesser General Public License
  * version 2.1 or later (LGPLv2.1+) WITHOUT ANY WARRANTY.
@@ -44,5 +46,23 @@ pe_resource_t *pe__create_clone_child(pe_resource_t *rsc,
 G_GNUC_INTERNAL
 void pe__force_anon(const char *standard, pe_resource_t *rsc, const char *rid,
                     pe_working_set_t *data_set);
+
+G_GNUC_INTERNAL
+gboolean unpack_remote_nodes(xmlNode *xml_resources, pe_working_set_t *data_set);
+
+G_GNUC_INTERNAL
+gboolean unpack_resources(xmlNode *xml_resources, pe_working_set_t *data_set);
+
+G_GNUC_INTERNAL
+gboolean unpack_config(xmlNode *config, pe_working_set_t *data_set);
+
+G_GNUC_INTERNAL
+gboolean unpack_nodes(xmlNode *xml_nodes, pe_working_set_t *data_set);
+
+G_GNUC_INTERNAL
+gboolean unpack_tags(xmlNode *xml_tags, pe_working_set_t *data_set);
+
+G_GNUC_INTERNAL
+gboolean unpack_status(xmlNode *status, pe_working_set_t *data_set);
 
 #endif  // PE_STATUS_PRIVATE__H
